War of the WaRnINGs Part the Seventh: First round through all four build types completed.
I will start working with Sergey to make warnings into errors THIS WEEK.

A small amount of warnings still remain that I've saved till the end. I will be posting to the list shortly for help with cleaning up those.

As far as tests, all builds compiled, unit tests (except the same broken one) passed, and all integration tests passed EXCEPT for 2: 1 is the broken darc one, but the other is make_rot_lib. I am not sure why, but the statistics are coming out slightly off in the log file. This may be because I changed the precision on some of the variables (int to Size, float and double to Real) per our coding conventions. If I need to change that back, please let me know.

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta/@54323 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: fd7ce353b47899e912ce71646fae9fc75cd96256

+++ rosetta_source/src/apps/pilot/stranges/min_pack_min.cc
@@ -232,7 +232,7 @@ void MinPackMinMover::apply (pose::Pose& pose ) {
 			// under option control for mutations that may push the complex away in minimization
 			movemap->set_jump(false);
 			for(Size ii = 1; ii <= jump_num_.size(); ++ii ){
-				if( jump_num_[ii] > pose.num_jump() )
+				if( jump_num_[ii] > (int)pose.num_jump() )
 					utility_exit_with_message( "Input declared a jump with too high a number, exiting... \n" );
 				else
 					movemap->set_jump( jump_num_[ii], true);
